requirements.txt
openai-whisper>=20231117
numpy>=1.24.0
torch>=2.0.0
tqdm>=4.65.0
